bul: Translate HoldMapTitle and RoomTypeNames
This commit aims to standardize most room types as "tokens" to reuse universally in all other occurrences. (Future commits will address this idea in short.) Translations are to the effect of "best possible", however, so conflicts may happen or weaknesses be found later after many translations.

HoldMapTitle applies for any recent item when the "Map" button is held, the rest applies to Teleport 2.0 and possibly a few others.